requirements.txt
Flask==1.0.2
fuzzywuzzy==0.17.0
Jinja2==2.10
pandas==0.23.4
paramiko==2.4.2
passlib==1.7.1
pdf2image==1.3.0
Pillow==5.3.0
pyodbc==4.0.25
pytesseract==0.2.6
requests==2.20.1
scikit-learn==0.20.2
urllib3==1.24.1
Werkzeug==0.14.1
XlsxWriter==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
WTForms==2.2.1